+++ internal/controller/suite_integration_test.go
@@ -14,6 +14,7 @@ import (
 	"testing"
 	"time"
 
+	"k8s.io/apimachinery/pkg/util/rand"
 	corev1 "k8s.io/api/core/v1"
 	networkingv1 "k8s.io/api/networking/v1"
 	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
@@ -248,7 +249,7 @@ func eventually(t *testing.T, timeout time.Duration, fn func() error) {
 // tests independent on the shared apiserver.
 func uniqueNS(t *testing.T, prefix string) string {
 	t.Helper()
-	return fmt.Sprintf("%s-%d", prefix, time.Now().UnixNano()%100000)
+	return prefix + "-" + rand.String(5)
 }
 
 func mustCreate(t *testing.T, obj client.Object) {